chore(mobile): split Timeline widget into smaller parts
